+++ src/spec/LoaderSpec.ts
@@ -1,11 +1,24 @@
 import * as ex from '@excalibur';
-import { ExcaliburMatchers, ensureImagesLoaded } from 'excalibur-jasmine';
+import { ExcaliburAsyncMatchers, ExcaliburMatchers, ensureImagesLoaded } from 'excalibur-jasmine';
 import { TestUtils } from './util/TestUtils';
 
 describe('A loader', () => {
   let engine: ex.Engine;
+
+  const reset = () => {
+    engine.stop();
+    engine.dispose();
+    engine = null;
+    (<any>window).devicePixelRatio = 1;
+    const playButton = document.getElementById('excalibur-play');
+    if (playButton) {
+      const body = playButton.parentNode.parentNode;
+      body.removeChild(playButton.parentNode);
+    }
+  };
   beforeEach(() => {
     jasmine.addMatchers(ExcaliburMatchers);
+    jasmine.addAsyncMatchers(ExcaliburAsyncMatchers);
     engine = TestUtils.engine();
   });
 
@@ -341,5 +354,33 @@ describe('A loader', () => {
       .catch(() => {
         fail();
       });
+  });
+
+  it('should not show the play button when suppressPlayButton is turned on', (done) => {
+    reset();
+    engine = TestUtils.engine({ suppressPlayButton: false });
+    engine.currentScene.add(
+      new ex.Actor({
+        pos: new ex.Vector(250, 250),
+        width: 20,
+        height: 20,
+        color: ex.Color.Red
+      })
+    );
+
+    const testClock = engine.clock as ex.TestClock;
+    const loader = new ex.Loader([new ex.ImageSource('src/spec/images/SpriteSpec/icon.png')]);
+    loader.suppressPlayButton = true;
+
+    TestUtils.runToReady(engine, loader).then(() => {
+      // With suppress play there is another 500 ms delay in engine load()
+      testClock.step(1);
+      engine.graphicsContext.flush();
+      expectAsync(engine.canvas)
+        .toEqualImage('src/spec/images/EngineSpec/engine-suppress-play.png')
+        .then(() => {
+          done();
+        });
+    });
   });
 });